Update some figure, plus use inherited Plotly figure and its add_scatter
Show the absence of evanescent components in the FFT spectrum.

Update heatmap with white empty marker to represent the PEC cylinder.

Use the add_scatter() methode of the Plotly Figure instead of add_trace
and inside the go.Scatter(). This way the graph_object do not have to
be imported in the main project, only the (inherited) Plotly Figure.

Calculate the Root-Mean-Square Deviation at the end of the script.