Move files from last semester in separate folder to serve as example
Move PWDExperience.py and FFTExperience.py with their data/ into folder
to separate and to serve as example or code base.

Move the new SimulationResultsPlotter.py as well which was used to plot
CST simulation results found as a txt in data/ folder.
However, would not commit that source file as it exceeds 42MB, so this
folder gets a gitignore file as well. SimulationResultsPlotter.py could
still be used as an example code too.