Create PW set and E-field T-matrix with MATLAB script
Create plane wave set with PlaneWaveSet.py, as it calculates the
equidistant sampling points along x, and calculates x- and y-components
of the wave vector "k" for each PW in the set. It also saves incident
angle of these plane waves in a .mat file for MATLAB simulation.

Saving the used MATLAB script files too, to be in the same place with the
related Python scripts:

mieScatt_singlePW.m is Oliver's script, used as example and base.

mieScatt_multiPW.m creates a structure and excites it with the
previously generated plane wave set. The correctly sampled E-field for
each PW has been saved into a (sample_size * pwset_size) shaped matrix,
called e_field_transfer_matrix. This is the base to generate the real
transfer_matrix of the structure, which will be used for the eigenmode
calculation.

I downloaded the MieScatter library too, but it's too big and does not
need version control, so it goes in a gitignore inside the MATLAB/ folder.
Same applies to the data used between Python and MATLAB in data/ folder.