Calculate eigenmodes and also filter and examine them
EigenModes.py can load transfer matrix and calculate its eigenvalues
and eigenvectors. It also sorts them by descending order of the
eigenvalues absolute value. These eigenvectors (and eigenvalues)
are the eigenmodes of the structure.

The eigenmodes can be examined manually with the mieScatt_ALLeigenPW.m
script, as it excites the structure with all the eigenmodes and
generates a figure containing the response at one line (y=-2) for
every eigenmode. This way the trivial solution can be filtered out
and the interesting ones can be selected manually.

These eigenmodes can also be filtered with eigenmode_filter() and with
the finetune of its input values. This can be used if some patterns
in the mieScatt_ALLeigenPW.m has been found (ex. 1.000 eigenvalues are
too perfect as they are the trivial solutions.)

Eigenvectors can be plotted on a polar figure to examine the excitation
of the eigenmodes. This can be used as a filtering methode too.

Finally, the "mieScatt_eigenPW.m" script can be used to check response
to the selected eigenmodes in a 2D space.