Skip some eigenmode and save result as a file too in ALLeigenPW.m
Firstly, this script was not reading the structure description from the
Python generated mat file as the rest, so I fixed it, and now it works.

I have added an option to skip some of the eigenmode, as they are
almost the same in the beginning. The skipped eigenmode results are
left as zeros in the e_fields matrix, so arbitrary section of the
eigenmodes can be simulated while not losing the results so far.
The waitbar and the remaining time calculator has been modified
accordingly to show real progress.

The final result can be saved in file too, so it can be plotted in Python
Plotly as well.