Upgrade EigenMode.py with userinput and parameterizable matlab script
The EigenModes.py script now asks the user if they want to plot the
eigen components and if they want to run the matlab script that simulates
and plots the eigenmodes.

It also asks the user if they want to simulate with the filtered and
suggested eigenmodes, or they can set the requested eigenmodes' indices
in the command line.

The mieScatt_eigenPW.m gets more inputs like the eval_x and eval_y
coordinate array to set the simulation area easily from Python.
It also gets a bool value if the user wants to simulate with the PEC
cylinder structure or set them as vacuum to simulate without them.
It also saves results in .txt for plotting in Python Plotly.

Matlab scripts received an output folder for the numerous outputs.

The matlab script runner in MatlabRunner.py can now measure how long
it took one script to run and writes it out on the console.
The matlab opener has been fixed, so the current directory is set
correctly.

Other small changes in TransferMatrixValidation.py and in gitignore.