Implement EigenModesPlotter.py to check eigenmode's E-field
This script only reads previously simulated data and plots them on a
heatmap. It can also calculate and plot the difference between two
simulation result, like an eigenmode simulation with and without the
scatterer structure in the way. This can be used as a measure how an
eigenmode is scattering-invariant.

Another script in this commit, the ALLEigenModePlotter.py similarly
can only plot results, but it only can plot the result of the
mieScatt_ALLeigenPW.m simulation, which is used to check every eigenmode
in a 1D line.

There was also a bug in the AskUserinputRecursively.py, where the
recursive calling was not returning with anything, making it ineffective
in case of a recursive calling.