Implement new PW simulation models to approximate realization
Since the plane wave is a strongly theoretical construct, the
approximation of its real life realization should be examined.
The MieScatter library can also simulate point sources, which is
a good model for an isotropic antenna.

The environment should be kept the way it is, so the PWD process can
still think in plane waves for inputs and outputs, but the simulation
translates the incident angle according to the model.

These models have been implemented in the mieScatt_multiPW.m and
mieScatt_somePW.m scripts so far. Some calculations can be made outside
and can be passed to the script as an argument, ex. the positions
of the point sources. It is connected to the plane wave set, so these
set-ups are placed in the PlaneWaveSet class as methods.

Three models have been implemented that translates the incident plane
wave to:
- point sources along a line at given distance from the origin (y const)
- point sources along a circle, whose middle is in the origin
- point sources as a phased array, whose array properties are defined,
but the phase shifts for each plane wave is calculated in MATLAB

The SimulationModelTester.py can set up each model and call the
mieScatt_somePW.m simulation to check how one or two plane waves
gets translated. It then plots the result on a heatmap, like the
EigenModePlotter.py

The said MATLAB scripts have also been modified in other places, like
- the load of the .mat files are made into a variables making them struct
- the mieScatt_somePW.m can be now evaluated for arbitrary points in
space, and can be used in 1D and 2D
- its output file name can also be defined in input argument
- the simulations can be made without structure
- the iterations have also been modified to not depend on i = i + 1