Bugfix the new simulation models and use them for Transfer Matrix
With the new models implemented, the TransferMatrix.py and the
TransferMatrixValidation.py scripts could be adapted to parametrize
and use these new models to create and validate T-matrices.

These and SimulationModelTester.py received an update on the progress
tracker and user input question, so they are easier to read.

Some bugs were found in the mieScatt_multiPW.m and mieScatt_somePW.m,
like some variable name and string were incorrect. Also, the matrix
multiplication was not working as intended for the phased array result
calculation, so the result was different when the field of each antenna
element was phase shifted in a loop one-by-one.
The MATLAB scripts are also received a progress tracker, so the wait
time can be estimated and the number of excitation can be seen.