Put scattering field specific plotting into function
The scattering field plotting was made with some fix steps
that could be place into a function in the MyPlotlyFigure.
This is the same situation as it was with the MATLAB styling.

These are the layout update with the anchoring axis scale,
the axis titling, and the plotting of the scatterer cylinders.