Enhance EigenModesPlotter.py with userinput and difference plotting
The script now ask the user to plot the eigenmode the usual way,
with the scatterers in the simulation field.
Then it asks if the user wants to calculate NRMSD and if yes,
if they want to plot this difference?

At the beginning, it also asks the indices, so it is not necessary
to write it into the script.