Use config files for MATLAB and set the model parameters once
Since the model parameters could have been set in any script,
some chaos emerged and the results were unpredictable.
The setting of the model parameters and the saving of them for
MATLAB only kept in the _TransferMatrix.py, along with the saving
of the PW set, which also suffered from the same effect.
The PW set is written into a pickle file, where it is reachable
for the other scripts.

MATLAB scripts are set up uniformly by configuration files, which
contains the simulation parameters (ex. eval point coordinates,
simulation model choice, PECorVacuum, PW indices, eigenvector).
This is then read into a "structure" in the MATLAB for easier
handling.

The waitbar has been removed, and the progress tracking in MATLAB
is moved to the terminal, where the same info is written out by
fprintf() funtion. This results in that now it can be seen while
it is run from terminal by MatlabRunner.

These are yet in the mieScatt_multiPW.m and the mieScatt_somePW.m

Also commenting and script progress tracker is enhanced.