Implement the simulation models in other MATLAB scripts as well
Use the configuration file usage for the simulations for the
mieScatt_eigenPW.m and mieScatt_ALLeigenPW.m as well as it
was introduced in the last commit for the other simulation scripts.

Create the different simulation model implementations based on
the other simulation scripts.
Also implement the new progress tracker, that writes in the
terminal instead of a waitbar.

Enhance mieScatt_ALLeigenPW.m to simulate the wave sources and
save each one to use for all the eigenmodes, instead of simulating
each one of them. This shortens the simulation from O(n*n) to O(2n).

Also save the scatterers' figure with the sampling points into
a pdf file.